refactor cache class
